+++ db/game_mutation.go
@@ -9,7 +9,7 @@ import (
 )
 
 // LockGameForUpdateTx locks the games row for the given id (SELECT … FOR UPDATE on supported drivers).
-func LockGameForUpdateTx(tx *gorm.DB, gameID uint) error {
+func LockGameForUpdateTx(tx *gorm.DB, gameID int64) error {
 	var row dao.Game
 	return tx.Clauses(clause.Locking{Strength: "UPDATE"}).
 		Where("id = ?", gameID).
@@ -17,7 +17,7 @@ func LockGameForUpdateTx(tx *gorm.DB, gameID uint) error {
 }
 
 // LoadGameByGameIDTx loads a full game graph using tx (same preloads as LoadGameByGameID).
-func LoadGameByGameIDTx(tx *gorm.DB, gameID uint) (*dao.Game, error) {
+func LoadGameByGameIDTx(tx *gorm.DB, gameID int64) (*dao.Game, error) {
 	var game dao.Game
 	err := preloadGameInfo(tx.Where("id = ?", gameID)).First(&game).Error
 	if err != nil {
@@ -28,7 +28,7 @@ func LoadGameByGameIDTx(tx *gorm.DB, gameID uint) (*dao.Game, error) {
 
 // WithGameMutationTx runs fn inside a single transaction after locking the game row.
 // All writes for this game in fn should use the same tx so commits are atomic across instances.
-func WithGameMutationTx(gameID uint, fn func(tx *gorm.DB, game *dao.Game) error) error {
+func WithGameMutationTx(gameID int64, fn func(tx *gorm.DB, game *dao.Game) error) error {
 	if gameID == 0 {
 		return fmt.Errorf("WithGameMutationTx: game id required")
 	}
