+++ lobby_server/roomclient/game_creator.go
@@ -14,7 +14,7 @@ type GameCreator struct {
 	Client proto.RoomServiceClient
 }
 
-func (c *GameCreator) CreateGameAndRun(players []types.PlayerAddress, gameType uint, completedMatchID int64) (uint, error) {
+func (c *GameCreator) CreateGameAndRun(players []types.PlayerAddress, gameType uint, completedMatchID int64) (int64, error) {
 	if c.Client == nil {
 		return 0, status.Error(codes.Unavailable, "room worker client not configured")
 	}
@@ -30,5 +30,5 @@ func (c *GameCreator) CreateGameAndRun(players []types.PlayerAddress, gameType u
 	if err != nil {
 		return 0, err
 	}
-	return uint(resp.GameId), nil
+	return resp.GameId, nil
 }
